requirements.txt
Flask==2.3.3
requests==2.31.0
Werkzeug==2.3.7
python-dotenv==1.0.0

